Add experimental TypeScript definitions, support passing subpaths as arrays
This focuses on type definitions for commonly-used Model methods.

To support useful generic types on Model and its methods, the signatures of Model methods will have to change somewhat. There are two options:

1. Allow the `subpath` parameter to Model methods to be an array - `model.set(['prop1', 'prop2'], val)` - which is the approach taken here.
2. Allow subpaths to instead be specified in a var-arg style - `model.set('prop1', 'prop2', val)`.

The first approach, array-path, is simpler to implement in the JS, which is why I took that approach for now. It plays well with the more complex methods like `model.on`, and it's easier to add fallback overloads for arbitrarily-many path segments.

However, the array-path approach does have a couple downsides:
- Worse autocomplete experience and type-check errors in some cases, mostly for array index path segments, since TypeScript seems to have a harder time with precise overload-matching.
- Very slightly more verbose.