Change path internals to use new _atSegments array where possible.
`this._splitPath(subpath)` used to string-concatenate `this._at` and `subpath`, then split on dots. It's called from every getter/setter, which is inefficient due to the concat+split and resulting string allocations.

By pre-splitting `_at` into `_atSegments`, we can skip a lot of that overhead, especially as more methods move towards taking a segment array instead of a path string.